requirements.txt
streamlit
smolagents
litellm
biopython
pytest